fix: remove unnecessary files
