+++ india_compliance/gst_india/report/gstr_1/gstr_1.py
@@ -29,7 +29,6 @@ from india_compliance.gst_india.report.hsn_wise_summary_of_outward_supplies.hsn_
     get_hsn_data,
     get_hsn_wise_json_data,
 )
-from india_compliance.gst_india.report.utils import get_company_gstin_number
 from india_compliance.gst_india.utils import (
     get_escaped_name,
     get_gst_accounts_by_type,
@@ -1855,6 +1854,39 @@ def get_rate_and_tax_details(row, gstin):
     return {"num": int(num), "itm_det": itm_det}
 
 
+def get_company_gstin_number(company, address=None, all_gstins=False):
+    gstin = ""
+    if address:
+        gstin = frappe.db.get_value("Address", address, "gstin")
+
+    if not gstin:
+        filters = [
+            ["is_your_company_address", "=", 1],
+            ["Dynamic Link", "link_doctype", "=", "Company"],
+            ["Dynamic Link", "link_name", "=", company],
+            ["Dynamic Link", "parenttype", "=", "Address"],
+            ["gstin", "!=", ""],
+        ]
+        gstin = frappe.get_all(
+            "Address",
+            filters=filters,
+            pluck="gstin",
+            order_by="is_primary_address desc",
+        )
+        if gstin and not all_gstins:
+            gstin = gstin[0]
+
+    if not gstin:
+        address = frappe.bold(address) if address else ""
+        frappe.throw(
+            _("Please set valid GSTIN No. in Company Address {} for company {}").format(
+                address, frappe.bold(company)
+            )
+        )
+
+    return gstin
+
+
 @frappe.whitelist()
 def download_json_file():
     """download json content in a file"""
@@ -1912,7 +1944,6 @@ def get_gstr1_excel(filters, data=None, columns=None):
     filename.extend([gstin, report_dict["fp"]])
 
     for type_of_business, abbr in report_types.items():
-
         if data_dict.get(type_of_business):
             continue
